agent/requirements.txt
aiohttp==3.8.4
ping3==4.0.4
psutil==5.9.4
dnspython==2.3.0
pydantic==1.10.7
pyyaml==6.0
toml==0.10.2
